+++ ui/src/components/RequestHistoryPanel.vue
@@ -5,11 +5,9 @@
             <li
                 @click="toggleExpand(snap)"
                 class="cursor-pointer p-1 my-3 rounded border"
-                v-for="(snap, ) in requestHistory"
+                v-for="snap in requestHistory"
                 :key="snap.id"
             >
-
-
                 <h2 class="font-semibold">
                     <span v-if="snap?.method" :class="getMethodColorClass(snap?.method)">{{
                         snap?.method.toUpperCase()
@@ -32,17 +30,18 @@
 
                 <div v-if="expandedIds.has(snap.id)" class="my-2 p-2">
                     <!-- <Btn @click.stop="retryRequest(snap)">Retry</Btn> -->
-                    <p >
+                    <p>
                         Status:
                         <span class="font-bold" :class="getStatusColorClass(snap.status)">{{ snap.status }}</span>
                     </p>
 
                     <div class="my-1 border-b">
                         <p class="">Query Params:</p>
-                        <p v-for="(value, key) in snap.queryParams">
-                        <pre
-                            class="bg-vs-pbg rounded overflow-x-auto p-2 text-xs max-h-100 my-4"
-                        ><code>{{ key }}: {{ value }}</code></pre></p>
+                        <div v-for="(value, key) in snap.queryParams">
+                            <pre
+                                class="bg-vs-pbg rounded overflow-x-auto p-2 text-xs max-h-100 my-4"
+                            ><code>{{ key }}: {{ value }}</code></pre>
+                        </div>
                     </div>
 
                     <div class="my-1 p-1 border-b">
@@ -68,15 +67,14 @@
 import { useRequestHistory } from '@/composables/useRequestHistory';
 import type { RequestSnapshot } from '@/types';
 import { getMethodColorClass, getStatusColorClass } from '@/utilities/dynamicColors';
-import { ref } from 'vue';
+import { onMounted, ref } from 'vue';
 // import Btn from '@/components/Common/Btn.vue';
 
-const { requestHistory } = useRequestHistory();
+const { requestHistory, loadHistory } = useRequestHistory();
 
 const expandedIds = ref(new Set<string>());
 
 function toggleExpand(snapshot: RequestSnapshot) {
-
     if (expandedIds.value.has(snapshot.id)) {
         expandedIds.value.delete(snapshot.id);
     } else {
@@ -84,9 +82,9 @@ function toggleExpand(snapshot: RequestSnapshot) {
     }
 }
 
-
-
-
+onMounted(() => {
+    loadHistory();
+});
 </script>
 
 <style scoped></style>